+++ src/EIDEProject.ts
@@ -45,7 +45,7 @@ import {
     ConfigMap, FileGroup,
     ProjectConfiguration, ProjectConfigData, WorkspaceConfiguration,
     CreateOptions,
-    ProjectConfigEvent, ProjectFileGroup, FileItem, EIDE_CONF_VERSION, ProjectTargetInfo, VirtualFolder, VirtualFile, CppConfigItem, ProjectBaseApi, ProjectType
+    ProjectConfigEvent, ProjectFileGroup, FileItem, EIDE_CONF_VERSION, ProjectTargetInfo, VirtualFolder, VirtualFile, CppConfigItem, ProjectBaseApi, ProjectType, BuilderConfigData
 } from './EIDETypeDefine';
 import { ToolchainName, IToolchian, ToolchainManager } from './ToolchainManager';
 import { GlobalEvent } from './GlobalEvents';
@@ -57,7 +57,7 @@ import { WebPanelManager } from './WebPanelManager';
 import { DependenceManager } from './DependenceManager';
 import * as platform from './Platform';
 import { IDebugConfigGenerator } from './DebugConfigGenerator';
-import { md5, copyObject, compareVersion } from './utility';
+import { md5, copyObject, compareVersion, isGccFamilyToolchain } from './utility';
 import { ResInstaller } from './ResInstaller';
 import {
     view_str$prompt$not_found_compiler, view_str$operation$name_can_not_be_blank,
@@ -1827,7 +1827,10 @@ export abstract class AbstractProject implements CustomConfigurationProvider, Pr
         return optFile;
     }
 
-    genLibsMakefileContent(makefileName: string): string | undefined {
+    /**
+     * @param makefile_repath a path relative from build output dir, like: '.lib/Makefile'
+    */
+    genLibsMakefileContent(makefile_repath: string): string | undefined {
 
         const fcfg = this.getLibsGeneratorCfgFile(true);
         if (!fcfg.IsFile())
@@ -1906,6 +1909,9 @@ export abstract class AbstractProject implements CustomConfigurationProvider, Pr
 
         for (const name in cfg) {
 
+            if (name.startsWith('$'))
+                continue; // skip internal vars
+
             let exprs: string[] = cfg[name];
             if (!Array.isArray(exprs)) continue;
 
@@ -1931,6 +1937,11 @@ export abstract class AbstractProject implements CustomConfigurationProvider, Pr
             }
         }
 
+        if (cfg['$AR_PATH'])
+            AR_PATH   = File.ToUnixPath(this.toAbsolutePath(cfg['$AR_PATH']));
+        if (cfg['$AR_CMD'])
+            AR_PARAMS = cfg['$AR_CMD'];
+
         // --------------------------
         // - gen makefile
         // --------------------------
@@ -1945,7 +1956,7 @@ export abstract class AbstractProject implements CustomConfigurationProvider, Pr
                 .replace('${out}', outname);
             let rule_tmp = `# ${libname}
 lib${libname}_OBJS += ${objs.join(AR_OBJ_SEP)}
-lib${libname}: $(lib${libname}_OBJS) ${makefileName}
+lib${libname}: $(lib${libname}_OBJS) ${makefile_repath}
 \t@echo -e $(COLOR_BLUE)-------------------------$(COLOR_END)
 \t@echo -e $(COLOR_BLUE)AR "${outname}"$(COLOR_END)
 \t@echo -e $(COLOR_BLUE)-------------------------$(COLOR_END)
@@ -1990,7 +2001,7 @@ $(OUT_DIR):
 
         mk_tmp = mk_tmp
             .replace('<LIB_TARGETS>', lib_rules.join('\n'))
-            .replace('<LIB_OUT_DIR>', '.lib');
+            .replace('<LIB_OUT_DIR>', NodePath.dirname(makefile_repath));
 
         return mk_tmp;
     }
@@ -3557,6 +3568,18 @@ class EIDEProject extends AbstractProject {
         }
     }
 
+    private _getCompilerIntrDefsForCpptools<T extends BuilderConfigData>(
+        toolchain: IToolchian, builderCfg: T, builderOpts: ICompileOptions): string[] {
+
+        if (['AC5', 'AC6'].includes(toolchain.name) || isGccFamilyToolchain(toolchain.name)) {
+            // we have provide a xxx-intr.h for cpptools,
+            // so return empty list.
+            return [];
+        } else {
+            return toolchain.getInternalDefines(builderCfg, builderOpts);
+        }
+    }
+
     private doUpdateCpptoolsConfig() {
 
         const builderOpts = this.getBuilderOptions();
@@ -3566,7 +3589,7 @@ class EIDEProject extends AbstractProject {
         // get project includes and defines
         const depMerge = prjConfig.GetAllMergeDep();
         const defMacros: string[] = ['__VSCODE_CPPTOOL']; // it's for internal force include header
-        const intrDefs = toolchain.getInternalDefines(<any>prjConfig.config.compileConfig, builderOpts);
+        const intrDefs = this._getCompilerIntrDefsForCpptools(toolchain, <any>prjConfig.config.compileConfig, builderOpts);
         const defLi = defMacros.concat(depMerge.defineList, intrDefs);
         depMerge.incList = depMerge.incList.concat(this.getSourceIncludeList()).map(p => this.ToAbsolutePath(p));